tecnologia/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-cached-authentication-middleware==0.2.2
django-qr-code==3.1.1
django-redis==5.3.0
django-redis-sessions==0.6.2
djangorestframework-jwt==1.11.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
reportlab==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
segno==1.5.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tabulate==0.9.0
typing_extensions==4.7.1
tzdata==2022.7
urllib3==1.26.14
yarl==1.9.3
